requirements.txt
accelerate==0.33.0
amqp==5.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diffusers==0.30.0
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
setuptools==72.2.0
six==1.16.0
sympy==1.13.1
tokenizers==0.19.1
torch==2.5.0.dev20240815
torchaudio==2.4.0.dev20240815
torchvision==0.20.0.dev20240815
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.20.0
